requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cloudinary==1.25.0
Django==3.2.4
djangorestframework==3.12.4
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
